+++ scripts/settings.js
@@ -92,7 +92,10 @@ const SETTINGS_GROUPS = {
       keys: ['wallCoverStandardThreshold', 'wallCoverAllowGreater', 'wallCoverGreaterThreshold'],
     },
   ],
-  Combat: [{ title: 'Flanking', keys: ['flankingSizeRule'] }],
+  Combat: [
+    { title: 'Combat Start', keys: ['raisePcShieldsWhenDefending', 'enrageBarbariansAtCombatStart'] },
+    { title: 'Flanking', keys: ['flankingSizeRule'] },
+  ],
   'A.V.S. Settings': [
     {
       title: 'A.V.S. Settings',